backend/requirements.txt
feedparser==6.0.11
supabase==2.9.1
python-dotenv==1.0.1
beautifulsoup4==4.12.3
httpx==0.27.2
fastapi==0.104.1
uvicorn[standard]==0.24.0
openai==1.54.0
